feat: handle response/exception in Buildings

+++ src/modules/buildings/buildings.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateBuildingDto } from './dto/create-building.dto';
 import { UpdateBuildingDto } from './dto/update-building.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -11,7 +11,7 @@ export class BuildingsService {
     @InjectRepository(Building)
     private buildingsRepository: Repository<Building>,
   ) {}
-  create(createBuildingDto: CreateBuildingDto) {
+  async create(createBuildingDto: CreateBuildingDto) {
     return this.buildingsRepository.save(createBuildingDto);
   }
 
@@ -19,15 +19,27 @@ export class BuildingsService {
     return this.buildingsRepository.find({ relations: ['locations'] });
   }
 
-  findOne(id: number) {
-    return this.buildingsRepository.findOne({ where: { id } });
+  async findOne(id: number) {
+    const building = await this.buildingsRepository.findOne({ where: { id } });
+      if (!building) {
+        throw new NotFoundException(`Building with ID ${id} not found`);
+      }
+      return this.buildingsRepository.findOne({ where: { id } });
   }
 
   update(id: number, updateBuildingDto: UpdateBuildingDto) {
-    return this.buildingsRepository.update(id, updateBuildingDto);;
+    const building = this.buildingsRepository.findOne({ where: { id } });
+    if (!building) {
+      throw new NotFoundException(`Building with ID ${id} not found`);
+    }
+    return this.buildingsRepository.update(id, updateBuildingDto);
   }
 
   remove(id: number) {
+    const building = this.buildingsRepository.findOne({ where: { id } });
+    if (!building) {
+      throw new NotFoundException(`Building with ID ${id} not found`);
+    }
     return this.buildingsRepository.delete(id);
   }
 }
